Migrate `nvtabular.inference` to `merlin.legacy.inference`
I don't think we can completely get rid of this code yet, since we're still relying on it to some extent for legacy PyTorch/T4R serving. As we sort out more of the remaining issues that are making it tricky to serve session-based models, we should be able to remove this entirely, but it still needs a home for now. It makes more sense for this code to live in Systems than NVT though, since that consolidates all the Triton code in one repo.