Remove unused legacy serving code
This code was copied from NVTabular with the intention of using it as a starting point for refactoring in the direction of an operator-based serving API (which now mostly exists), but the remaining unused code was never removed. This PR removes it an effort to avoid confusing users with multiple versions of the legacy serving code they could (attempt to) use.