Replace flake8 with ruff (which is equivalent but faster)
`ruff` is a very fast Python linter written in Rust that covers the flake8 rules in addition to a bunch of other popular linters. This sets up a config for it that covers the existing flake8 config, and adds additional linters while generating minimal errors with the existing code base. Since `pylint` is much slower than `ruff` (and I suspect slower than `flake8` too), this config executes the `ruff` checks first in order to fail fast.