Adjust Tensorflow ensemble tests to use model with correct input shapes
It turns out you can't build a model where the input shape is only the batch size with the Tensorflow Functional or Sequential APIs—you can only do it by sub-classing `tf.keras.Model`. So this reworks the example models used in the tests to have that shape, and adjust the test expectations to use the resulting output name.