Downgrade unused columns error to a warning (#306)
Creating columns you don't use is still a bad idea, but it used to crash Triton when we were using Triton ensembles and now is mostly just going to cause the user headaches when they need to supply information that isn't actually used. Downgrading it to a warning will let people do this but still signal that it's a bad idea.